+++ packages/iov-bns-governance/src/proposals.ts
@@ -15,7 +15,6 @@ export enum ProposalType {
   ReleaseGuaranteeFunds,
   DistributeFunds,
   AmendProtocol,
-  TreasurySend,
 }
 
 interface CommonProposalOptions {
@@ -80,14 +79,6 @@ export interface AmendProtocolOptions extends CommonProposalOptions {
   readonly text: string;
 }
 
-export interface TreasurySendOptions extends CommonProposalOptions {
-  readonly type: ProposalType.TreasurySend;
-  readonly recipients: readonly {
-    readonly address: Address;
-    readonly amount: Amount;
-  }[];
-}
-
 export type ProposalOptions =
   | AddCommitteeMemberOptions
   | RemoveCommitteeMemberOptions
@@ -97,5 +88,4 @@ export type ProposalOptions =
   | RemoveValidatorOptions
   | ReleaseGuaranteeFundsOptions
   | DistributeFundsOptions
-  | AmendProtocolOptions
-  | TreasurySendOptions;
+  | AmendProtocolOptions;